IJPL-175177 Fix the New Element positioning in RD
In RemDev, because the popup is created on the backend,
the input event doesn't have a reference to an ActionButton.
It seems to have a reference to some fake backend component,
though, and passing that does the job just fine.j

It also works in the monolith because nothing has effectively
changed from the previous implementation.
This code is only invoked if the action is triggered by a mouse
event, and that is only technically possible through an action button,
so the check wasn't strictly necessary.

GitOrigin-RevId: 284cb5986a7459d85f1e574223502254c76fadb4