KTIJ-33249 [kotlin] Introduce a cache for `getKotlinSubclassObjectsByNameFilter` call
This call is not cheap, since it has to pull all subclassed objects from the indices.
If there are a lot of unresolved references in the file, this can be quite expensive.

See the KDoc for `SUBCLASS_OBJECTS_SYMBOLS_CACHE` about why this cache is safe
regarding `KaSession`'s validity.

GitOrigin-RevId: 156ce24d413a3739876055d6d509db64e29f18bd