[terminal] IJPL-179295 Ensure that input events are handled in the same order on backend as they are sent from frontend
Use `SendChannel` for this purpose.
Just calling `sendInputEvent` method in the right order does not guarantee the same order on backend.
To guarantee it, we have to wait until `sendInputEvent` method completes (the event is handled on backend) and only then send the new one.
But it is not acceptable in case of typing in RD: the latency would be very noticeable.
We should not wait for handling of the previous event before sending the next one, we should send it immediately.
And hopefully, `SendChannel` acts exactly like that, and also it provides the order guarantees.


(cherry picked from commit f3180a974c6bf8af6d12bd0ffe162d99ed95a31c)

IJ-CR-156212

GitOrigin-RevId: 47aaf8f20d202ec4ad69d8f9c16ccff4e18997ce