[RDCT] IJPL-178633: Fix Code Style settings page missing its content
`CodeStyleSchemesConfigurable` is written under an assumption that methods are always called in this order:
1. `buildConfigurables`
2. `createComponent`
3. `disposeUIResources`

So, panels (`myPanels`) are created in the first method, accessed in the second, and cleared in the third one.
But in Remote Development, unlike Monolith, `buildConfigurables` method is called only once for each project (and for extension point updates). This led to the Code Style page having its content only on the first show.
This commit fixes that, calling `buildConfigurables` method when needed.


(cherry picked from commit 32f0b13b67f30610ef976e0b71284268a2d56944)

IJ-CR-156422

GitOrigin-RevId: 0c10af9ddcf7e9266de489ac3fcff798c7eaa639