[remote dev] don't return 'null' from 'getService' for per-client services if the session is disposed (IJPL-168959)
It may happen that a session is disposed, but some code from an async task still tries to take a service from it. In that case, ClientSessionImpl.doGetService shouldn't return null because it'll cause NullPointerException, but delegate to the base implementation, which will throw an instance of CancellationException in such a case and the task will be cancelled.

This fixes sporadic failures in RestoringStateTest.

GitOrigin-RevId: 2189c0fd2c6997e07759e539f07bd5cdaa5da26c