[terminal] IJPL-179295 Ensure that input events are sent in the right order
Launching a coroutine with `limitedParallelism` does not help with that. Because it doesn't guarantee that all launched coroutines will be executed sequentially. So, after we finish awaiting the channel, the events may be sent in a different way.
Use flow with `tryEmit` that guarantees that events will be collected sequentially.

GitOrigin-RevId: 78d378364f20c355d880be33dd3bfb52625a6c5b