KTIJ-33266 [kotlin] Mark candidates from the current package as already imported in `AbstractImportCandidatesProvider`
Technically, this fix is not 100% correct, since
adding an explicit import for a declaration
can actually improve its priority in the resolve,
and allow to get rid of some red code.

So this is viewed as a temporary solution
before KTIJ-33214 is fixed.

`extensionWrongTypeParam2_class` testdata is ignored in K1
because it tries to import the `Foo` class
from the same package and fails to do that
because it is already imported.

^KTIJ-33266 Fixed

(cherry picked from commit 5729ed9da79be1637c987abd3364337f601f193d)

IJ-CR-158539

GitOrigin-RevId: d2ffed12ece19e280c5bee29c34a57e99cd0605b