[jps build] store the set of enabled instrumenters and rebuild modules if it's changed (IJPL-163450)
A new interface JvmClassFileInstrumenter is introduced to allow rebuilding modules when an instrumenter becomes enabled or disabled.
For now, it's implemented by UI Forms instrumenter only, and the new behavior is disabled by default. We can implement this interface by other instruments as well (see IJPL-171493), and after that enable the new behavior.

GitOrigin-RevId: dcf818decf0197a267e55a0fda9bdf54a1a6d046