[plugin model] properly disable plugin modules with duplicating names (IJ-CR-155237)
If two plugins include modules with the same name, only one of these plugins can be loaded. However, before the module from the other plugin was added to PluginSet::enabledModules, because its name was registered in enabledModuleV2Ids. And it was necessary to work around this problem by skipping such modules in ClassLoaderConfigurator.configureModule (4c58def12fd8117104). Now such modules aren't included in the 'enabledModules' set, so the workaround is removed.

GitOrigin-RevId: a323a61083742001a54df5210306c3e689bbd6a7