[terminal] IJPL-179449 Cleanup: remove migration of deprecated useShellPrompt setting
It would be nice to also remove the setting from the `TerminalOptionsProvider`, but we can't do it immediately because it is used in the JSON settings: IJPL-156187.
So, `com.intellij.ide.settings.json.JsonSettingsTest` fails in case of removing.
We need to investigate it separately.

(cherry picked from commit acfcc54bb3bcca9074d101f68d1c1ed4d16c8939)

IJ-CR-157330

GitOrigin-RevId: 23aabc88b70e304b5b9e028cae3d545507f6f330