IJPL-175886 Fix Alt+O in the Hierarchy tool window
The base action didn't work because the exporter
was not included in the data context.
The toolbar action did work because it uses a subclass
that gets the exporter into the constructor.
It's not clear why it was implemented that way,
this code is very old. But at least the fix is very simple.

GitOrigin-RevId: dc9e03a392dd1ed24af7b028d3d689df4b4de215