IJPL-179770 Fix `DbContext.threadLocal` being reset on EDT due to coroutine context propagation
Problem:
`updateDbInTheEventDispatchThread` used `withContext(Dispatchers.EDT)` and `log.collect`, which triggered `DbSource.ContextElement.updateThreadContext`.
At that point, `DbContext.threadBoundOrNull` was still null, so it was stored as `oldState`.
When `restoreThreadContext` ran, it restored this null value, effectively discarding the `DbContext` we had just set in `updateDbInTheEventDispatchThread`.

To ensure that `threadLocal` is not overwritten by `restoreThreadContext` we need to avoid using coroutines continuation (`withContext` and `collect`).
Instead of that, `runInEdt` can be used.

`transactor().log.collect` is not really necessary here as long as we set context with a proper `DbSource`.
DB snapshot is going to be updated by `DbSource.ContextElement` (`updateThreadContext` and `restoreThreadContext`) on coroutine's continuation.

GitOrigin-RevId: 002e63de973500fe3d40df14070314061506549f