IJPL-175250 Add file to indexable files filter before applying pushers
Pushers may schedule file for indexing via FileBasedIndexImpl.requestReindex which adds them to ChangedFilesCollector.myDirtyFiles so at that point they should already be in the indexable files filter otherwise they are added to myDirtyFilesWithoutProject.

Note that it looks like a bug that scanning adds files to ChangedFilesCollector. But in general pushers should be allowed to check indexable files filter

GitOrigin-RevId: 2427e8eab9a4646d7bfa9684017549e3bb11f849