[threading] Emulate `invokeAndWait` under modal progress
Without the future, this code would rely on having all modal events being dispatched before the exit of modal progress. Apparently there is no such guarantee -- we can have modal events that are executed with `Modality.current()` which is different from the events' modality. Therefore, we rollback to the previous solution with `invokeAndWait` here. It would not cause deadlock anyway, because events are generated inside modal progress

GitOrigin-RevId: 82476771da4d1db1ae6ffb7b944f5953a5fd8348