[terminal] IJPL-180098 Add terminal engine option to the terminal settings
Use checkbox to specify the engine. Checkbox value is composed from the `terminal.new.ui` and `terminal.new.ui.reworked` registry key values.
Setting the new value in this checkbox updates the registry key values.
So, it is just a new visualization for existing settings and there is no need to update the clients.

Also, I had to remove the logic executed when Gen1 terminal is enabled: it included FUS reporting, local logging and triggering feedback notification. Since we plan to deprecate Gen1 terminal, let's remove this logic because it isn't valuable anymore.

GitOrigin-RevId: 2fa74bb7a1edd3d944896539080376ef6ecab230