[plugin model] fix problems with loading required and embedded modules in the core plugin (IJPL-180209)
The main module of the core plugin is always put in the first place by 'toCoreAwareComparator', because all other modules have implicit dependency on it. So the approach used for required modules in other plugins (add edges from the main module to all required modules) doesn't work for the core plugin: the main module was processed before required content modules became enabled, so the core plugin was disabled and the IDE fails to start. Now this logic is switched off for the core plugin, and a separate code checks that all required modules are enabled and throws EssentialPluginMissingException if some module isn't.

The build scripts are changed to include 'embedded' content modules in the core plugin to a JAR file in the lib directory.

GitOrigin-RevId: 4ef0bbce73b9f4695924263e7df62006f8533273