[java-inspections] ObviousNullCheckInspection: disable for methods that don't return the parameter
Even if the return value is not used, it's subject for IgnoreResultOfCallInspection to report it (as we consider only pure methods). It could sometimes be suppressed (e.g., inside assertThrows), in which case it's desired not to report anything.
Fixes IDEA-368213 False positive and code-breaking autofix for ObviousNullCheck


(cherry picked from commit 0ab2fd64b8a8f8599920cffad76fc830790f8269)

IJ-CR-156868

GitOrigin-RevId: cda30602e0c2af93f21a33501f7ad855b89fda1c