[driver] register intellij.driver.* modules as content modules in Performance Testing plugin
This is needed to properly load them as part of the plugin in the frontend process (JetBrains Client). Before, 'intellij.driver.client' was incorrectly loaded as part of the core plugin because intellij.platform.frontend.split had a dependency on it. This change is also needed to get rid of artificial plugin aliases (see IJPL-177155), because IdeaUltimateAdditionalBuildTest.jetBrainsClientForIdeaBuildSuccessfully starts to fail when dependencies for plugin modules are migrated.

GitOrigin-RevId: e4837464cb89ff4db661fde4f72881ca19be651d