[terminal] IJPL-179949 Make Paste action scroll to cursor
When Paste action is called by shortcut, its keystrokes are consumed the platform. So, our logic of scrolling in `TerminalEventsHandlerImpl` is not called.
It should be an action's responsibility to scroll to the cursor if its UX requires that.


(cherry picked from commit f0ab4826fb749110f9d226e24149620dbf79f2d8)

IJ-CR-157016

GitOrigin-RevId: 270f8909a161d2fa09becd1daec0223fc4100bb5