IJPL-180322 Fix TreeState selection restore for flattened nodes
Since introducing flattened nodes, we implemented
restoring the expanded state for them, but not the selected state.

To fix, we need to reuse the existing matching mechanism
in applySelectedTo(). But because selection was previously
implemented by prioritizing user object over serialized data,
we need to modify the matcher to account for that.
The new matcher API allows to check what kind of match it was,
and if it was the low-priority one, we now also have the API
to reset the matcher back and save the new state in case
we need to fall back to it.

While we're at it, reuse the same cache for selection.
This should fix potential freezes when refreshing large
flat trees with large selection (think Select All + Refresh).

GitOrigin-RevId: 355933aac931e663f460a827173039b25183b6f9