IJPL-177414 Send resize events to the terminal when its font changes
To do this, the font options listener had to be extracted from
TerminalUiUtils and moved to the frontend, as it's no longer
local to the editor. The API had to be made public-@Internal
so that it can be accessed from the frontend.

As a side effect, the font isn't applied anymore to other
terminal implementations besides the reworked one.
It's OK as they're going to be deprecated anyway.

GitOrigin-RevId: d65ebd63cfc1a2e50c1db8897b32912c5bdbe687