IJPL-160286 Synchronize document and entity
Before the synchronization worked only one way - from document to entity.
Now it can synchronize both ways. It's required to support documents synchronization in RD mode where events can happen on the frontend and have to be processed by backend's document.

Both `DefaultAdEntityProvider` and `FrontendAdEntityProvider` also have to subscribe to document changes to enable synchronization through Rhizome.
Before only `DefaultAdEntityProvider` did that, so typing on the frontend side wasn't sent to the backend at all (it worked because of the patch engine, which is disabled now)

GitOrigin-RevId: a9ff84f55941d36f32868d4ca1322b987873adfe