IJPL-160286 Fix "failed to reconsider shared transaction" - "Required value was null"
The exception was thrown when during rebase when inside one `shared` block, we tried to execute a second `shared` block.
`DbMutableDocument.edit` is called in a shared scope and uses `maybeShared` inside, which causes an exception.
It seems that we should execute `DbMutableDocument.edit` only under a local `change` block and rely on the "mutate" implementation to decide which scope to use.

GitOrigin-RevId: 812bd9be9ef0884e83f9d1c42bef73f737739d8a