IJPL-164422 Editor grid mode refactoring: extract an interface
The logic of the character grid mode is being spread over
various pieces of code and it becomes harder to follow.
Character width calculation is in one place, terminal size
calculation is in another, etc.
Now that we're about to optimize soft wraps and size calculations,
it's going to add even more to this mess.
To avoid this, extract the relevant logic into a class
and expose it via an interface.

GitOrigin-RevId: 16411688015b7f3961e1bdda41e543070c60cfd6