[rdct] IJPL-169760: restricts editor-based fallback to `IdeFrame`
After debugging, I can conclude that the only component that needs an editor-based fallback is basically `IdeFrame`. The focus (on the backend) will be set to an `IdeFrame` when on the frontend the focus is set to a frontend-specific component. Since we could not (?) build a map from a frontend component to a backend-component, we just fall back to the editor.

GitOrigin-RevId: f5201abacc516175286e833a2211998793dc6c3a