IJPL-164422 Optimized soft wraps for cell grid editors
When outputting large text into a terminal, it spends about 30-40%
of the time calculating soft wraps. Most lines, however, don't have
inlays, and because characters are aligned to cells, no font metrics
is required to calculate soft wraps. We only need to know the number
of columns and for every character wither it's a double-width one
or not.

Introduce a special soft wrap mode that automatically kicks in
as soon as the grid mode is enabled and there are no inlays in the range.

To avoid creating an iterator just to check whether there are inlays,
extract inlay calculation from the constructor.

To avoid performance issues caused by repetitive codePointAt() invocations,
make a copy of the whole thing using toString() instead. This will create garbage,
of course, but the performance benefit seems to be very well worth it.

GitOrigin-RevId: d34e7522e32691faf70e5dd3e921dfb4856bbac7