IJPL-180831 Optimized preferred width for cell grid editors
Besides soft wraps, this was the most time-consuming thing.
It would recompute the preferred width on every document
trimming. And it's not necessary at all because for a grid editor
the preferred width is fully determined by the number of columns
and the width of a single cell. The only case when it's not so
is when the actual size isn't initialized yet.

GitOrigin-RevId: 7dfd58b845bdcd0b8a1ed596b8cc42a389923706