IJPL-177172 Eel+GeneralCommandLine: run processes using IJent only if executable AND workdir belong to that IJent
The previous code had a problem. The Database plugin runs `java.exe` from the IDE distribution inside the WSL project. That behavior was intentional.

The commit 2a0baa9fd8eab4843fb05e1f7a1fb8b5e23ddefc introduced new logic: `GeneralCommandLine` runs on the custom Eel if the executable or the working directory are on a non-local Eel. This behavior led to the regression because the Database plugin lost the ability to run `java.exe` for WSL classes.

This commit changes the logic: the command runs inside the Eel only if both the executable path and the working directory are on the same non-local Eel.

NB. At this moment, the only production-ready implementation of Eel is for WSL. This code won't work with the upcoming Eel on Docker, but that's another problem that should be fixed separately.


(cherry picked from commit 8534c048b775b8df82fe0b18d624c46125e1d8ba)

IJ-CR-157461

GitOrigin-RevId: b4c1c3c7e3f7a87243138ea7ea42c61d40aea933