IJPL-164422 Move DWC strategy to CharacterGrid
Since we're concentrating all things grid in one place,
it makes sense to have the strategy there.
It's already only used there. The only reason why it was
outside was the default implementation, which was
closely coupled to EditorView internals. But it was not
actually used anywhere, so it doesn't seem to be a good reason
to write bad code.

GitOrigin-RevId: 78fe1fdc6615a533115b4439d787c0414d61ab8e