IJPL-177414 Default line spacing for the terminal: 1.0
As agreed, we're setting the default value to 1.0.
Previously it was inherited from the console / editor settings,
that is, 1.2.

A special case is when the user set this value to something
different from the default. In that case we keep the value
set by the user.

Also overwrite the default in the default constructor,
so that the serialization mechanism compares the current
value to the correct default, when determining whether
it should be saved or not.

(cherry picked from commit a2588d924bcf26ad056024400da9b8bd13348856)

IJ-CR-156671

GitOrigin-RevId: c1fc7cd3e382db34e494bdddd9a00be8fa5d70cc