IJPL-180885 Fix terminal cursor update for a trimmed line
If the cursor is on the first line, and it was partially trimmed
(e.g. a very long line being output), then its column index
is not equal to the column index in the document anymore.
This causes a lot of extra spaces to be added to the end
of the line to display the cursor somewhere far beyond
the end.

Fix by keeping track of the characters trimmed from the first
line and by adjusting the cursor position accordingly.
Make sure this new counter is persisted as a part of the state.

Add a test to check both the fix and state dump/restore.

GitOrigin-RevId: 051518e8f3ae8a8726c627a9c2de55a08c109b53