[build scripts] put 'embedded' content modules to separate JARs if they are used by the frontend process (IJPL-179421)
If we put content modules with 'loading=embedded' in the main plugin JAR, they cannot be reused by the frontend process, because this would lead to unrelated entries on the classpath, and tests which verifies integrity of the embedded frontend will fail. Now such modules are put in a separate *-frontend.jar in the 'lib' directory. They are still loaded by the plugin's main classloader.

KDoc in ModuleLoadingRule is changed accordingly (actually, 'embedded' modules from the core plugin was put to a separate JAR even before this change).

An explicit dependency on 'intellij.platform.backend' module is added to intellij.pwa.core and intellij.qodana.sarif to prevent their extraction to separate JARs (it's not necessary since these modules don't make sense in the frontend process anyway).

GitOrigin-RevId: 1f82b295a49a03a17a074de46246815e9a6716ec