Always use compiler version as the max value for api level when known
For figuring out the API level, we currently use the minimum of compiler version and the stdlib version module's library dependencies. This is an expensive operation and it can be avoided when we know the answer beforehand. My understanding is that we know the answer when compiler version is set explicitly (this is at least true for Gradle import case). In all other cases we can still use the old logic.

closes https://github.com/JetBrains/intellij-community/pull/2956

relates to KTIJ-33413

(cherry picked from commit f2703c1d54de266aed7444733871df109ee610ee)

Merge-request: IJ-MR-157187
Merged-by: Aleksei Cherepanov <aleksei.cherepanov@jetbrains.com>
GitOrigin-RevId: 85d6908b7886e2174479273d91296d870aa77433
Change-Id: I614c02c09b8d577df3872451354b1b7016b3541a