[JEWEL-695] Fix ScrollableContainer layout, add content alignment
This fixes the ScrollableContainer to allow it to "fit content" on both
the main axis, and the cross axis. This means that it won't try to take
as much space as possible even when not set to fillMax*(), but rather it
will only use as much space as it needs.

It also makes the scrollable container's content have a BoxScope as its
receiver, so that users can align the content when it's smaller than the
scrollable container itself (e.g., center it vertically if shorter than
the viewport size).

The Scrollbars sample in the showcase has been updated accordingly.

GitOrigin-RevId: ee51c75b5ab1b2901c9615bf9e3015b35badce22