[json + remdev] IJPL-176240 Add json.backend dependency for any third-party plugin that declares dependency on the json plugin id
This is required for the case when external developers have their plugin implemented using plugin model v2, meaning that dependency on the JSON plugin id is not enough to have access to classes from the backend module. This module contains json schema-related code, which is the most popular extension point, so we have to make sure users are able to access them in runtime. With the change monolith-compatible third party plugins do not require any additional changes to be able to work with the updated json plugin layout

GitOrigin-RevId: 8ba8b392e62c9caff9a86d3bb847c9b55aafbf8f