IJPL-180546 Repaint affected components when the gradient position changes
We have gradient painting through ProjectWindowCustomizerService.paint
in several components, but we don't have code that would repaint
those components when the gradient is updated. For example,
when the project widget is added / removed to the toolbar
or when its position changes. This caused painting artifacts.

To fix this for good, introduce a project service that keeps track
of the current gradient position. It's updated by the project widget
itself, so it's always up to date. Then use this service to repaint
the affected components.

Bridging the two together turned out to be harder than it seems
because when the component is shown, the project is often not known
yet. So we keep the list of all affected components, and carefully
clean it up when the components are hidden, to avoid leaks.
Then, when the gradient position changes for some project,
we go through the list and update the affected components.

Note that some updates are redundant, for example, when
the tool bar is updated, all components are temporarily removed
from it, which triggers a show/hide event pair.
This doesn't seem to actually affect the look in practice
because these events happen very fast one after another.


(cherry picked from commit f333c9b4d80f954089656566974cb3a56a0711ad)

IJ-CR-157878

GitOrigin-RevId: bd3dccc404213082a5dfd7b40cc2e4db56c25da4