[formatter] IJPL-165316 fix cached are CodeStyleSettings recomputed using invalid VirtualFile
Computed CodeStyleSettings are cached by their URL, and so the cached value may outlive the original VirtualFile for which it was requested.
If a new VirtualFile appears at a previously cached URL, and CodeStyleSettings are requested for it, we get problems.

Previously, in such situations, a new CodeStyleCachedValueProvider with the new VirtualFile would eventually be created. However, that is not enough; a CachedValue stores its original provider and uses that to recompute its value if needed, so the recomputation would still use the old VirtualFile. The new provider is used by the stability checker and correctly warned us that the context has changed.

To fix, this commit adds more thorough cleanup in this situation: the entire CachedValue is replaced.

GitOrigin-RevId: 5055e932753569fcca1373ae462c4959702f7d23