[JEWEL] Add modifiers to Markdown block renderer APIs
This allows implementers of custom renderers to pass in a modifier to
the default renderer. This is particularly convenient when the
customization solely consists in aspects expressible by a modifier.

Also in this, we fix the Markdown code blocks not looking disabled when
the rest of the rendered Markdown does (JEWEL-465), and tweaks the
default editor text style line height to better match the IDE's.