[terminal] IJPL-182482 Implement terminal frontend typing latency FUS
The latency is measured from the Swing input event
to the moment when the resulting event is written to
the session's input channel, as that's where the frontend
part ends.

Measuring consists of two phases: Swing input event processing,
which is easy because it's single-threaded and is put in a try-finally
block; and terminal input event processing, which is trickier
because of the coroutines and the channel. Thankfully,
the channel provides API to ensure that a "resource" is always
closed: either the send operation fails, or the element is received,
or an "undelivered element" callback is fired. In order for it to work,
the channel also must be canceled (not just closed) eventually.
This allows us to keep an event-to-FUS-activity map and ensure
that it's always cleaned.

Input events are assigned global sequential IDs without corresponding
session IDs or anything like that because at the time the Swing event
arrives, it's not known yet which terminal session it belongs to,
or even if it belongs to any at all. So we just measure the time it takes
for an event to arrive to any session, if at all. It shouldn't be
an issue because it's impossible to have parallel input activities.

The latency is reported in microseconds, because it's very often
less than one millisecond.

GitOrigin-RevId: a31bdfe60d9e32fb16d0391d855a5e26011bf8fb