[terminal] IJPL-182482 Implement terminal backend typing latency
This process is not single-threaded, but it involves only
one thread transfer: from the receiving coroutine
to the terminal thread, which uses an executor
that is never shut down, so it's pretty much a guarantee
that for every byte array there will be at least an attempt
to write it. Using this fact, we can store a byte array to activity
map and remove activities when the byte arrays are written.
Using a try-finally block there ensures that the activity is eventually
finished.

The activity ID is transferred along with the event,
to ensure that the same ID is used on the frontend and on the backend.

GitOrigin-RevId: 66587a5cca59053dab1e4bf2bee9bd098fe10637