Eel+WSL: don't throw errors on double deregistration of EelDescriptor in EelNioBridgeService
That behavior led to sporadical test failures in Rider. Those failures were not easy to reproduce.

Presumably, they could have been caused by the following race:
* Some test called `com.intellij.ide.starter.extended.targets.Target.getInstance` that calls `registerIjentNioFs`.
* After a while, `deregister` was called by a coroutine inside `registerIjentNioFs`.
* After a while, `IjentWslNioFsToggleStrategy.unregisterAll` tried to remove it again.

The old function is temporarily kept in case if it is used inside some feature branch.

The new function doesn't throw exceptions and returns a boolean instead. Is called `unregister`, because it is impossible to overload a function by changing the return type.

GitOrigin-RevId: dbd19dd2b210d39c0805df64215c1ebaca4dc98a