[terminal] IJPL-182482 Implement terminal backend output latency FUS
This is much harder than the input latency.

The output has a lifecycle of several phases.

The first phase: chars are read from the TTY connector.
This happens on the terminal emulator thread,
and we get a whole array of chars with a single timestamp.

The second phase: chars are processed by the emulator
at its own pace. This also happens on the terminal emulator
thread, so at least we can track this processing in a synchronous
manner. Some characters arrive into the text buffer,
but not all of them, as some control characters don't affect
the buffer. So we need to be careful to figure out
the char indices that actually made their way into the buffer.
The text buffer is updated under the text buffer lock,
but not all processing happens under this lock.

The third phase: the text buffer is collected (scrapped).
This doesn't happen on the terminal emulator thread,
but it happens under the same text buffer lock.
At this point an event is created, and we must carefully
match it to character indices.

The fourth phase: the event is collected by the frontend.
Finally we can measure the latency. But the latencies of the first
and the last characters can be different if they were read
at different times. So we report two latencies.

A dangerous part is that we need to ensure that events
are always collected. The current implementation guarantees
it. If during startup or shutdown an event or two are not collected,
it's not much of an issue as there's no global state here,
unlike the input latency measuring. The entire activity
is disposed along with the session.

Because, unlike input, it's possible to have parallel outputs,
every event is reported with the session ID.

(cherry picked from commit 50c2626163e03d905a4b62762276fda59b450452)

IJ-CR-160384

GitOrigin-RevId: 961f222ee3f1ad15879af2087ba7cdfa007d639f