[terminal] IJPL-182482 Implement terminal frontend output latency FUS
The hardest part here is to understand when the received data
is actually displayed. The problem is that it may not be displayed
at all, for example, if the editor is hidden or scrolled away.

Using the hooks added to the editor, we can make an assumption:
if the editor is showing, and there was a repaint request,
then it's very likely to be repainted soon. In that case we wait
for the repaint before reporting the latency.

If there's no repaint request, then we report the latency immediately,
adding a boolean flag to the event to distinguish between the cases.

Because events are received and applied as whole,
there are no separate events for the first and last characters.
Instead, we report both character indices in one event.
So for two backend events we have one frontend event.

(cherry picked from commit df72687dc3ef31bece6eeaf7cd624dc18c462d6e)

IJ-CR-160384

GitOrigin-RevId: ae6f6527f332cd31f20735238a9748249aa27bc5