[terminal] IJPL-182482 Fix the terminal tests to account for FUS activities
There are two ways to fix this: either to imitate the real
scenario or by relaxing the state checks in the FUS activities
themselves. Otherwise tests fail because the FUS activity
is invoked in the wrong state (e.g., characters make their
way into the text buffer without being read first).

Let's fix it on the test side to avoid making changes
to production that are only needed for tests.

GitOrigin-RevId: 922f947f3c745de620112ec85b6e6932116c8a2b