IJPL-181532 Report terminal font settings to FUS
All three numeric defaults are now exposed
as internal properties. This is to avoid duplicating
them, even if they're referring to an existing
value like FontPreferences.DEFAULT_FONT_SIZE.
Because later we may wish to change that to something
terminal-specific, and we'd like to keep this logic in one place.

GitOrigin-RevId: edf776eec9e04b9f787f3093dd105c0bbeb09e57