IJPL-180678 Add null value debouncing to the project gradient offset
When the main toolbar is updated, all widgets are removed first
and then re-added. This causes the gradient offset value to be
reset to null and then back to its normal value. Sometimes it causes
visual blinking.

Fix by debouncing the offset, but only if it's actually null.
Setting it to null is uncommon anyway, and normally can only happen
when the user actually removed the project widget.
In that case we're totally OK with the gradient reacting
100 milliseconds later.


(cherry picked from commit 4ee82594e65b3f7abe689274a2f663a0fdc0b227)

IJ-CR-157878

GitOrigin-RevId: c4c08419b3b9029fe4484ac0184758b3da91e045