PY-78919, KTNB-954: Fix color scheme change
Use `DelegateColorScheme` because it's the only way to update it (see `EditorImpl.MyColorSchemeDelegate#updateGlobalScheme`). Also, `DelegateColorScheme` isn't unwrapped in `setColorsScheme`; only `MyColorSchemeDelegate` is unwrapped there, so the comment was adjusted to reflect it.

To solve a chicken-and-the-egg problem, unwrapping is now performed in `DefaultNotebookEditorAppearance`: we need non-delegate there because a delegated scheme may contain old color value.

GitOrigin-RevId: 29c0776a810e6465d8b3b70e929889b005e21300