Eel/IJent: Remove ability to terminate Windows processes
Initially, there was a common contract for Posix and Windows processes:
* kill: SIGKILL on Posix, TerminateProcess on Windows.
* interrupt: SIGINT on Posix, Ctrl+C on Windows.
* terminate: SIGTERM on Posix, a hack with ProcessExit on Windows.

The latter one is suspicious: it brings a non-existent feature to Windows processes, which, by the way, doesn't work reliably. Also, there was no demand for that feature: it was borrowed for granted from FSD.

Although java.lang.Process provide methods `destroy` and `destroyForcibly`, they do the same on Windows.

This commits removes the `terminate` method for Windows processes, to conform java.lang.Process abilities and to not provide unneeded functions.

GitOrigin-RevId: 3e0a9b90470e0110dd9afd2d90e0dc2f4e73ef0f