KTIJ-30680 [kotlin] Handle `COMPONENT_FUNCTION_AMBIGUITY` diagnostic in auto-imports in K2 Mode
This is done because K2 compiler is more prone to producing `COMPONENT_FUNCTION_AMBIGUITY`
diagnostics compared to K1 compiler.
For example, it currently does it for every flexible type,
see KT-76253.

Since this can be problematic for users who have lots of flexible types in their codebase
(for example, users of Reactor, see KTIJ-33579), it's better to also handle
this diagnostic for now, and to re-evaluate this approach in the future.

^KTIJ-30680 Fixed

(cherry picked from commit 6bc4bb2ab56868868c7d106fcad35adf816dad5c)

IJ-CR-158539

GitOrigin-RevId: b74d819d0c0b167793635fb36719f87fc2663490