[collab/gh] Add missing checks for setting the default account (IJPL-73040)
What could happen is that one would set an account to default and
remove that account in the same transaction.
In that case, the default account would still be set and not checked.

We should probably remove or revise the notification completely,
since many things tend to work fine without an explicit default account.

#IJPL-73040 Fixed

GitOrigin-RevId: 40421b83468d0957d9ff1a238e52d3ad9f7d13c4