[github] Add a fake ghost user for GHE servers (IJPL-174769)
Didn't check every single usage, but in an ideal world,
this user should never leak into an API request.

Erroneous usages of the FAKE ghost user should be found when reported
and replaced with `null`, because in this case it's apparently
possible not to have a user assigned at all.

To be abundantly clear: this FAKE ghost user replaces github.com/ghost
in case one does not exist on the server.

GitOrigin-RevId: 6a1f48e6751992463cbe5c43b740ef2ab0f999fa