[Jupyter Debugger] PY-80007 Debugger needs 3 additional calls of F9 to stop after facing an exception #PY-80007 Fixed
Fixed detection that this file is actually a single Jupiter cell.
Previously, we checked whether this file was in the IPython cache.
However, not only executed cells were in the cache, but generally all files that were needed to execute cells: library files, user .py files that were imported.

Therefore, now it is checked that this file is in the cache and does not have a physical .py file.

More details: In IPython core 8.x, this means that the cache recorded the creation time of this file as None, for earlier versions, the cache had a separate map, _ipython_cache, which stored only Jupiter cells.

GitOrigin-RevId: cd97409192987cf063cbf2cd9bc03acf61dad513